Moved exceptions and logger to the root of the package scope. This is to make these modules more accessible to the upcoming importer. Added some error logging where exceptions are also raised. Logger no longer assumes the file name. Sub-packages (exporter, importer) will define this name upon init. Performed some code cleanup in exporter/geometry.py.
